refactor: pokemonの型ファイルを移動
