+++ app/routes/__.poc_.pokemons/types/pokemon.ts
